First-Task-Flask/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.23.4
opencv-contrib-python==4.6.0.66
Pillow==9.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
